refactor: 기존 용병 페이지를 route path에 연결한 컴포넌트로 이동
