+++ src/App.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import Router from './Router/Router';
 import { DefaultTemplate } from '@/components';
