+++ src/pages/index.ts
@@ -1,7 +1,5 @@
-export { default as MatchesPage } from './MatchesPage';
 export { default as Matches } from './Matches/Matches';
 export { default as Hires } from './Hires/Hires';
-export { default as HiresPage } from './HiresPage';
 export { default as Teams } from './Teams/Teams';
 export { default as Users } from './Users/Users';
 export { default as Login } from './Login/Login';
